requirements.txt
python-telegram-bot>=21.0,<23.0
openai>=1.0.0,<2.3.0
profanityfilter>=2.1.0
pytest>=7.0.0
